+++ app/src/main/java/com/neqabty/presentation/ui/notifications/NotificationsFragment.kt
@@ -124,7 +124,7 @@ class NotificationsFragment : BaseFragment(), Injectable {
     fun loadNotifications(typeID: Int) {
         llSuperProgressbar.visibility = View.VISIBLE
         binding.tvError.visibility = View.GONE
-        notificationsViewModel.getNotifications(typeID, 0, PreferencesHelper(requireContext()).user.toInt())
+        notificationsViewModel.getNotifications(typeID, 0, PreferencesHelper(requireContext()).user)
     }
 
     //region
